Move notebook assistant DTOs to common layer
Extract INotebookCellDTO, INotebookContextDTO, and NotebookCellType
from the API layer to workbench/common/positron for proper sharing
between contrib and API layers.
