+++ src/vs/workbench/api/browser/positron/mainThreadNotebookFeatures.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { extHostNamedCustomer, IExtHostContext } from '../../../services/extensions/common/extHostCustomers.js';
-import { MainPositronContext, MainThreadNotebookFeaturesShape, INotebookContextDTO, INotebookCellDTO, INotebookCellOutputDTO } from '../../common/positron/extHost.positron.protocol.js';
-import { NotebookCellType } from '../../common/positron/extHostTypes.positron.js';
+import { MainPositronContext, MainThreadNotebookFeaturesShape, INotebookCellOutputDTO } from '../../common/positron/extHost.positron.protocol.js';
+import { INotebookCellDTO, INotebookContextDTO, NotebookCellType } from '../../../common/positron/notebookAssistant.js';
 import { IPositronNotebookService } from '../../../contrib/positronNotebook/browser/positronNotebookService.js';
 import { IPositronNotebookInstance, NotebookOperationType } from '../../../contrib/positronNotebook/browser/IPositronNotebookInstance.js';
 import { IPositronNotebookCell, CellSelectionStatus, IPositronNotebookCodeCell } from '../../../contrib/positronNotebook/browser/PositronNotebookCells/IPositronNotebookCell.js';
