+++ src/vs/workbench/api/common/positron/extHostNotebookFeatures.ts
@@ -5,7 +5,7 @@
 
 import * as extHostProtocol from './extHost.positron.protocol.js';
 import type * as positron from 'positron';
-import { NotebookCellType } from './extHostTypes.positron.js';
+import { INotebookContextDTO, NotebookCellType } from '../../../common/positron/notebookAssistant.js';
 
 /**
  * Extension host implementation of notebook features.
@@ -22,7 +22,7 @@ export class ExtHostNotebookFeatures implements extHostProtocol.ExtHostNotebookF
 	 * Gets the context information for the currently active notebook.
 	 * @returns The notebook context DTO, or undefined if no notebook is active.
 	 */
-	async getActiveNotebookContext(): Promise<extHostProtocol.INotebookContextDTO | undefined> {
+	async getActiveNotebookContext(): Promise<INotebookContextDTO | undefined> {
 		return this._proxy.$getActiveNotebookContext();
 	}
