+++ src/vs/workbench/contrib/positronNotebook/browser/IPositronNotebookInstance.ts
@@ -17,7 +17,7 @@ import { IPositronNotebookEditor } from './IPositronNotebookEditor.js';
 import { IHoverManager } from '../../../../platform/hover/browser/hoverManager.js';
 import { IPositronNotebookContribution } from './positronNotebookExtensions.js';
 import { IInstantiationService } from '../../../../platform/instantiation/common/instantiation.js';
-import { INotebookContextDTO } from '../../../api/common/positron/extHost.positron.protocol.js';
+import { INotebookContextDTO } from '../../../common/positron/notebookAssistant.js';
 
 /**
  * Represents a deletion sentinel - a temporary placeholder shown where cells were deleted.
